requirements.txt
PySide6>=6.5
Pillow>=10.0
qtawesome>=1.3
pillow-heif>=0.15
